refactor: explicit barrel file, move methods around
