--- packages/utils/src/constants.ts
@@ -1,2 +1,0 @@
-export const SAFE_TIMERS_SYMBOL: unique symbol = Symbol('vitest:SAFE_TIMERS')
-export const SAFE_COLORS_SYMBOL: unique symbol = Symbol('vitest:SAFE_COLORS')
