+++ packages/utils/src/display.ts
@@ -1,5 +1,10 @@
 // since this is already part of Vitest via Chai, we can just reuse it without increasing the size of bundle
 import * as loupe from 'loupe'
+import type { PrettyFormatOptions } from '@vitest/pretty-format'
+import {
+  format as prettyFormat,
+  plugins as prettyFormatPlugins,
+} from '@vitest/pretty-format'
 
 type Inspect = (value: unknown, options: Options) => string
 interface Options {
@@ -17,6 +22,57 @@ interface Options {
 }
 
 type LoupeOptions = Partial<Options>
+
+const {
+  AsymmetricMatcher,
+  DOMCollection,
+  DOMElement,
+  Immutable,
+  ReactElement,
+  ReactTestComponent,
+} = prettyFormatPlugins
+
+const PLUGINS = [
+  ReactTestComponent,
+  ReactElement,
+  DOMElement,
+  DOMCollection,
+  Immutable,
+  AsymmetricMatcher,
+]
+
+export function stringify(
+  object: unknown,
+  maxDepth = 10,
+  { maxLength, ...options }: PrettyFormatOptions & { maxLength?: number } = {},
+): string {
+  const MAX_LENGTH = maxLength ?? 10000
+  let result
+
+  try {
+    result = prettyFormat(object, {
+      maxDepth,
+      escapeString: false,
+      // min: true,
+      plugins: PLUGINS,
+      ...options,
+    })
+  }
+  catch {
+    result = prettyFormat(object, {
+      callToJSON: false,
+      maxDepth,
+      escapeString: false,
+      // min: true,
+      plugins: PLUGINS,
+      ...options,
+    })
+  }
+
+  return result.length >= MAX_LENGTH && maxDepth > 1
+    ? stringify(object, Math.floor(maxDepth / 2))
+    : result
+}
 
 const formatRegExp = /%[sdjifoOc%]/g
